Change publishers scp to use list for sources
Use a list of source entries to be combined by the module code to
provide a more readable yaml when configuring jobs with a large number
of source entries.

Provide backwards compatible support with a deprecation notice to users
to switch from a standard string to using a list instead.

Change-Id: If6b48062e245754f2f732da0b99c312245c727c1