Update patch set 5
Patch Set 5: Code-Review-1

(3 comments)

Hey sorry for taking so long to get to this. This is potentially going to be really useful for us, I do have a slight concern on how parallel will react to a busy system but that's testable on our end.

Just a few nits to fix but I think this is a great patch.

Patch-set: 5
Label: Code-Review=-1