Update patch set 1
Patch Set 1:

Hi there, and thanks for contributing!

How much of this is common with the shining panda configuration?

If there are sections that are the same I'd recommend pulling them out into the helpers.py file which contains functions that can be reused if multiple configuration areas require the same settings. You can create new function for this common config.

Patch-set: 1
Label: Verified=0