Update patch set 11
Patch Set 11:

Patch 11 is just a rebase. It does not take care of any of the concerns from patch 10. I'll propose a patch 12 next to try to address some of the feedback in patch 10.

Brandon, I hope you don't mind I'd like to try to help push this patch along as I think it will be very useful.

Patch-set: 11
Label: Verified=0