Update patch set 1
Patch Set 1:

Like I said in https://storyboard.openstack.org/#!/story/2000438, there are more changes needed in order to have the CMake builder working with versions 2.x of the Jenkins plugin. In particular, dropping obsolete arguments and adding some new ones. When/if done, this will mean that the builder will not work anymore with old versions of the plugin, is this acceptable ?

Patch-set: 1
Label: Verified=0