Update patch set 21
Patch Set 21: Code-Review-1

(1 comment)

There is something not quite right still here. I ran this with the comparison script where I enabled the deep merge behaviour by default for the new configs and I was seeing that for the openstack jobs that the timeout specified under the template '{pipeline}-tempest-dsvm-neutron-pg-full{job-suffix}' was getting overridden by the default of 30 from the timeout defined in the macros.

It looks like the problem is if you have a macro that results to module and you use the module explicitly elsewhere, you end up having one overriding the other at expansion time.


i.e.:

 project-config/jenkins/jobs/devstack-gate.yaml
 - job-template:
     name: '{pipeline}-tempest-dsvm-neutron-pg-full{job-suffix}'
     node: '{node}'
    
     wrappers:
       - timeout:
           timeout: 125
           fail: true
       - timestamps

 project-config/jenkins/jobs/macros.yaml
 - wrapper:
     name: build-timeout
     wrappers:
       - timeout:
           timeout: '{timeout}'
           timeout-var: 'BUILD_TIMEOUT'
           fail: true

 project-config/jenkins/jobs/defaults.yaml
 - defaults:
     name: global
     description: ...
     project-type: freestyle
     concurrent: true

     wrappers:
       - build-timeout:
           timeout: 30
       - timestamps


This might simply be a limitation due to how macros are currently resolved right now, but we should call that out in the documentation of this behaviour since it's likely to catch people by surprise.

What we need is also an expansion of the macros first, for both defaults and templates, and then an application of deep merge of the defaults to make sure we can spot anywhere that we have something that a macro would be overriding if applied.

I'll see if I can add a test in the new few days that illustrates the problem by adding the second test mentioned.

Patch-set: 21
Label: Code-Review=-1