Update patch set 13
Patch Set 13:

Thanks for rebasing it. happy to see confirmation the error in build was not caused by me.

About the warn then that is only going to be there in the case users disabled it - for connections to servers where the certificate is know you will not get it.

I thought this was a good compromise to be security aware.

Patch-set: 13
Reviewer: Gerrit User 9614 <9614@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0