Update patch set 1
Patch Set 1:

I strongly suggest that this one only merges in close cooperation with the infra team to monitor any fallout from this. We had to revert this kind of patch twice since it broke all tox changes, so let's be careful the third time and only merge when people are around to monitor and can keep an eye for problems - instead of us figuring out what's broken out of the blue.

Patch-set: 1
Label: Verified=0