Update patch set 22
Patch Set 22:

I think this tickles the problem a little more explicitly. Now this may need to be treated as a separate bug, where we should be doing macro expansion of everything before generating the XML instead of both at the same time.

If that is the case, then we should document it as a limitation of the current deep merge, and state that you must be careful to use the same macro in the defaults as anywhere you need to use the module.

Otherwise we should look to fix that behaviour before merging this change.

Patch-set: 22
Label: Verified=0