Update patch set 26
Patch Set 26:

So, compare-xml-old and compare-xml-new has a twofold purpose:

1) Check for backward compatibility when new code is added
2) Run the code against a more complex collection of jobs/templates/macros/defaults than can be expected from the unit tests.

In hiding this rather complex functionality behind a config option, the deep merge capability does not get well exercised with complex job configurations. Which is why it's throwing up some issues once I try to run it against it.

Obviously we don't want to break backwards behaviour and cause surprises for end users without doing a proper major release where this can be expected, but at the same time we need a way to exercise code that is not enabled by default, with a more sophisticated set of input data.

Hence my suggestion that we create a another tox env, that explicitly uses config options not enabled by default that result in noticeable changes in behaviour to ensure that such code executes without throwing exceptions, and doesn't get broken due to not catching all possible configurations with some unit tests.

Patch-set: 26