Update patch set 1
Patch Set 1:

Why can't we do both? I.e. give the 'pip' command in the README.rst (which I assume is what ends up on the pypi website?), but under the "Developers" header, give the command for installing from source (maybe after the `git clone` command). Similarly in 'installation.rst'.

Also note that one issue is that the online documentation refers to the `git HEAD` version, which can lead to confusion. If the common case is that people will install from pypi, then I'd say it should match the git tag of the latest release on pypi. And then for developers, have an option for viewing the HEAD version of the documentation.

Patch-set: 1
Label: Verified=0