Update patch set 2
Patch Set 2:

(1 comment)

Is adding it to pep8 really a good idea given it relies on the external site to be available to complete successfully?

I'm thinking the resulting job should be similar to the jenkins-job-builder-compare-xml (now that it prints a sensible warning when changes occur), in that we only warn that there is a problem with the links instead of blocking as it may be just a connection issue and not that the link itself is wrong. Also true that links can become invalid over time without being incorrect when originally added.

Patch-set: 2
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0