Update patch set 1
Patch Set 1: Code-Review+1

ok, presumably setting the XML on the older versions won't cause issues?

Jenkins does appear to be quite good at ignoring XML it doesn't understand, but it's still a good idea to check that doesn't mean the job itself is not created or updated with the older verison of the plugin once this setting lands.

Otherwise this looks good to land, just need to test the above before I +2 unless someone else gets to that before me.

Patch-set: 1
Label: Code-Review=+1