Update patch set 6
Patch Set 6:

The comment is intentional, but there might be some projects that slipped through the cracks.  The intent is that we do not enforce on hacking checks because the hacking project represents the opinions of some of the nova developers.  It does not represent universally accepted practices, and is often far too unstable for practical use.

The point of not using it is that in general, none of the things it checks are actually important for infra projects.  If people are spending time reviewing for things that are not enforced, we should politely ask them to please stop.  It is not worth contributor or reviewer time to make minor unimportant changes.

In other words, we are trying to support a culture where people review changes for functionally important things, not for nitpicks.

Patch-set: 6