Update patch set 3
Patch Set 3:

FYI, I added everyone in the "jenkins-job-builder-core" group to this review but there are 16 other commits in this topic that each build on this--I decided not to add the same group to every commit since I thought that might be annoying.

Something I am curious about is whether or not the "jenkins-job-builder-compare-xml" test time of ~15m is within the standard deviation for that particular test? I am accustomed to seeing times in the range of ~5m, but I'm hoping there is occasionally some kind of backend-load-based slowness that can be expected (otherwise it seems I'll need to find a performance regression, which might be easier if I break this commit down into smaller pieces).

Patch-set: 3
Reviewer: Gerrit User 13487 <13487@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0