Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

doc doesn't show up as intended. While fixing that, probably worth tweaking the code slight as well to help out the user a little more and avoid putting in values that only appear in the XML and not in the UI that they might be referring to.

I think it's probably better to use what the UI provides for the values to be specified by the user into JJB and then translate that to the Java/XML, with some exception checking, or leave them out altogether and tell the user to look at the plugin for the correct non-default alternatives.

Generally lean towards the former which just means a few additions in the code around this to complete.

Patch-set: 2
Label: Code-Review=-1