Update patch set 5
Patch Set 5:

I think this brings it up to date with the folder enhancements made to python-jenkins, so it should now be feasible to merge.

However, need to actually test against a project jenkins instance with a folders plugin installed just to be certain the basic upload/modify/delete and delete-all work as expected when working with multiple jobs but different folders.

Additionally there is another change up to support creation of folders where there may need to be some thought before we merge this one about how things are going to work together. See https://review.openstack.org/189697

I suspect we may need another improvement to python-jenkins, for a check folder exists or some way to capture a list of available folders on the jenkins instance as well before attempting to upload to provide some sensible error to users should a target folder not exist.

Depends on how careful we want to be. It may be reasonable to depend on the users to create folder, but if we're managing both, it would be bad to not notice if we're creating a job instead of a folder and then try to create more jobs inside of it, or fail to create them in the right order for success.

Patch-set: 5
Label: Verified=0