Update patch set 6
Patch Set 6:

(1 comment)

The callback would be there to handle cases that a simple mapping doesn't, both conditional (options that are only valid if another is specified first) and nested (sub options to process on items if defined), which exist for quite a few of the components.

Whether it's better to handle checking if an argument is Required, Optional or has specific valid values via the callback mechanism, might be more efficient code wise given default values would be most common, or having custom objects to describe optional/required/must-match as the default value is unclear.

I'd even be tempted to see if having a custom class MappingValue of which Required/Optional/MustMatch be either instances or subclasses so a single check would be sufficient to determine if we have a default value or require more complex processing.


Using Thanh's suggestion, lets just focus on the minimal required to resolve for this patch and we can loop around with a subsequent patch to expand the capabilities further.

Should be enough here for us to think about anyway.

Patch-set: 6