Update patch set 1
Patch Set 1:

Why was this patch abandoned?

The Gerrit workflow is to "git commit --amend" your patch and upload a new patchset updating the existing review. It seems your new upload depends on this one. What you need to do is merge your other patch back into this patch so that they are a single commit. (If you open gitk you should have exactly 1 commit off of origin/master)

Gerrit uses the Change-Id field to determine if your commit is an update to a patchset. My recommendation for you at the moment would be to run:

    git rebase -i HEAD~2

and select change the "Raise Missing..." commit to "fixup" which and leave this commit as "pick" this will tell Gerrit to squash your 2nd commit into this one. Hope this helps in which case you can push it back online. (Don't forget to restore this patch first).

Patch-set: 1