Use system default encoding if current is None
There is possibilities that stream encoding is None rather than
raising an AttributeError exception.

In this case, the stream will be wrap again and again until recursive
limit exceeded when accessing stream's attribute.

For example, testing following job configuration will lead to this
failure.

The unit test is not added because the result xml it too large to be
included in git repo.

Test case:
```
- job-template:
    name: '{name}-{a}-{b}-{c}-{d}'

- project:
    name: project-name
    a:
      - 0
      - 1
      - 2
      - 3
      - 4
      - 5
      - 6
      - 7
      - 8
      - 9
    b:
      - 0
      - 1
      - 2
      - 3
      - 4
      - 5
      - 6
      - 7
      - 8
      - 9
    c:
      - 0
      - 1
      - 2
      - 3
      - 4
      - 5
      - 6
      - 7
      - 8
      - 9
    d:
      - 0
      - 1
      - 2
      - 3
      - 4
      - 5
      - 6
      - 7
      - 8
      - 9
    jobs:
        - '{name}-{a}-{b}-{c}-{d}'
```

Change-Id: I4eecc0b9e28bbc48987565252e7374734816d541
Signed-off-by: Hao Liu <hliu@redhat.com>