Complete support of Yaml Axis Plugin
The plugin uses a different matrix execution strategy. I have refactored
the part of code assuming there is only one strategy.  The previous one
is known as 'default' and this change is back compatible.

Drop an error whenever more than one strategy is used.

Support exclude rules from either a file or inlined YAML (note in the
later case the YAML will need to be a literal string).

Change-Id: I18e0fc7cb525fba3b772bae10e81d8dfce7298cb