Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I'm pretty certain the docstring has a typo, could you check please.

If it is incorrect, I think we do need to fix it up before merging since the meaning difference is pretty significant.

Second query I have is purely informational in nature, no need to change anything in the patch for this one, just want to know if we need to do something subsequently.

Patch-set: 1