Update patch set 1
Patch Set 1:

(1 comment)

Seems like it would have been a good idea to include an explanation as to why it was important to support use of 'null' in this case, as you can just leave the value empty in the yaml and it would result in the same output.

Since the pattern of "<dict>.get(<name>, '')" is used throughout the code it would be important to determine whether we need to maybe support converting None at the parser level or do a mass change to the codebase to use this pattern.

Perhaps in this case it just wasn't clean you could leave the yaml value empty?

Patch-set: 1
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0