Update patch set 1
Patch Set 1:

We were using a tool that generated yaml, which we then fed to jjb. This was a bit of breakage I saw when using the generated yaml.

I'd be happy to work on another patch that removes keys with None values at the parser level, so that components don't have to worry about this and we support null values, and revert this change at the same time.

Patch-set: 1
Reviewer: Gerrit User 12100 <12100@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0