Update patch set 4
Patch Set 4: Code-Review+2

Just tweaked it to move the example added without excludes to be in the tests as well so that we can be sure that both methods are checked.

I think it definitely shows how much neater and clearer it is to maintain this information within one project definition instead of needing multiples.

Patch-set: 4
Label: Code-Review=+2