Update patch set 3
Patch Set 3:

(1 comment)

Only query is if there is a particular reason not to use '.get()' with a default value of an empty list here.

Can we sensibly assume it's ok not to provide a list of projects without checking other options, or are you thinking there are too many possibilities and we should leave it up to the user to check whether they need it with what options they have given?

Patch-set: 3
Label: Verified=0