Update patch set 1
Patch Set 1: Code-Review+1

Hi, I can confirm that this is a bug. This bug only surfaces when used with Python 3. This does not happen in Python 2.7 as 2.7 automatically encodes that implicitly. Additionally, this bug does not happen in case JJB fetches configuration from its cache but only when JJB fetches config from Jenkins - regardless of whether or not the job was originally created bt JJB.

I can also confirm that this patch fixes the issue although I'm not entirely sure whether its the right way to do so.

Steps to reproduces:
1. Install JJB via pip3
2. Run jenkins_jobs on a new job
3. Clear JJB cache - one way to do this is installing JJB in a new container
4. Try to update the job created in step 2

Patch-set: 1
Label: Code-Review=+1