Clarify exception message for missed trigger properties
Currently, if we have a job configured for the
`comment-added-event` trigger with missed required
properties the error during its build is a bit
confusing.

For example, for config:
```
...
    trigger-on:
      - patchset-created-event
      - comment-added-event
...
```

The error in Jenkins logs looks like:
```
File ".../jenkins_jobs/modules/triggers.py", line 115,
in gerrit_handle_legacy_configuration
('approval-category', data['trigger-approval-category']),
KeyError: 'trigger-approval-category'
```
That leeds to additional requirement to debug the code
of the `triggers.py` file.

I propose to raise an exception in such situation, with
the message:
```
The comment-added-event trigger requires which approval
category and value you want to trigger the job.
It should be specified by the approval-category
and approval-value properties.
```

Change-Id: I33de93126d6ddbb6283ab32526c97657749e9208