Update patch set 6
Patch Set 6:

> Uploaded patch set 6.

patch set 6 addresses all the comments. Though I haven't used an helper function as suggested but xml parent tag to determine when we're using a publisher or a builder. It fixes also the regression introduced in the publish-over-ssh publisher as it doesn't have an extra top delegate element compared to the builder.

Patch-set: 6
Label: Verified=0