Fix CFP configuration for >=2.10.0
In CFP plugin version >=2.10.0 the JJB code completely breaks for Maven
settings due to the new support for freeform cfp id names. There's no
easy way to fix this for 2.10.0 support and not break backwards
compatability so I decided to make 'file' the default as that was the
behaviour before cfp support was added. To get JJB to load a cfp id
instead of a filename we must now pass the optional parameter
'settings-type' in the YAML file.

Users using cfp will now need to set 'settigns-type: cfp' in their YAML
files if they need to pull in CFP settings. All other settings file
configuration will work as it did in the past.

Change-Id: I650d485de6ecc1d2ef8fad6580c0c315fadde168
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>