Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Comment suggests that it may leave a system in an inconsistent state, without really explaining why.

I suspect it's really the use of 'sudo' and lack of using '--user' as the default install option that would leave the system in an inconsistent state. Possibly some version of pip introduced 

Assuming that this is the intention I think that the doc should be changed to use 'pip install .' without the sudo, the note that for older versions of pip (or unpatched versions as Ubuntu adds a custom default here) that you can use the '--user' option to enable use of a per user site location.

Patch-set: 1
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1