Update patch set 3
Patch Set 3:

Each cache is specific to a master, so if you run against jenkins on a different host you are loading a different cache.

Issues:

1. Want to avoid corrupting cache in case of exit during cache file write.

2. Want to run and update against a particular jenkins master and prevent accidentally running a second instance against the same master loading it unnecessarily by updating the same jobs.

3. Want to prevent multiple instances of JJB running at the same time on the same host. (maybe I misunderstood this one?)


Issue 1 would be covered by using a temporary file to write the cache to and then moving. Think I'll do that with a separate patch to keep the discussion on these separated.

Issue 2 seems like a lockfile would be needed.

Issue 3 seems very much a specific user requirement that is not necessarily valid for all users and would be better accomplished through whatever configuration management tool is being used to take out a global lock file using something like:

 (
     flock -n 9 || exit 1
     # run JJB
 ) 9>/var/lock/mylockfile

Patch-set: 3