Update patch set 2
Patch Set 2: Code-Review+2

(2 comments)

I'm in two minds whether we should do the clean up of the code as part of this one or as a separate patch.

For now leaning towards a follow on patch, but possibly it'd be good to use the correct exception for the added code and use '()' around the list comprehension + conditional to avoid the trailing '\' for line continuation.

Since the existing code in the same function also needs the same applied to it, we could do it all in one.

Patch-set: 2
Label: Code-Review=+2