Update patch set 6
Patch Set 6: Code-Review-1

(4 comments)

Hopefully I'll get to validate the output in a test jenkins before approving this, it's just 'odd' enough for the builder that it should be given a second set of eyeballs.

Just a few minor nits around the docstring so that it'll appear consistent in the documentation and fix the example to display correctly.

Patch-set: 6
Label: Code-Review=-1