Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I think the code is a little over verbose, but main reason I think this shouldn't be merged as is, it looks very like it uses the jclouds backend, and from a quick read of the upstream plugin, it seems like they have abandoned that backend and replaced it with a custom one.


Btw, the data formatting of the existing jclouds wrapper happens to allow for cleaner looking code, something to bear in mind when updating this to follow the more recent plugin.

Using the same format, I suspect that the manual-template would be a valid input to the jclouds plugin, except it's missing from ours, and it would be possible to modify both functions to use a common one with only the wrapper tags being different, instead of needing separate.

That would certainly be the way I would favour accomplishing this should you be on the older version of the plugin and wish support to be included for that. There are a few examples around, just grep the codebase for 'registry.get_plugin_info', and you can use the same pattern to handle plugins that change XML format between versions.

Patch-set: 3
Label: Code-Review=-1