Update patch set 3
Patch Set 3:

Sorry I should have been a little more verbose, reviewing too many things at the same time.

Main thing is I think the code needs to be brought inline with the predominate style as when your familiar with the existing style.

The other points where mainly observation and things I think are worth looking at before merging.

a) I think the input format of the existing jclouds one allows for cleaner code and could be adopted for this one as well.

b) The plugin changelog suggests that this code may already be out of date with respect to the XML now generated. Which may mean that landing this as is means we might have to support a change in XML which usually results in checking the version of the plugin installed and then branching to the different XML formats. Potentially using the same input layout format as the jclouds plugin would allow us to support the old format through a common function and supporting the new format as the main content of this module.

Just a possibility of saving ourselves from some messy code, which I think warrants checking before approving as well.

Patch-set: 3