Update patch set 3
Patch Set 3:

> Main thing is I think the code needs to be brought inline with the
> predominate style as when your familiar with the existing style.

I do agree that you have a point on this one.

> a) I think the input format of the existing jclouds one allows for cleaner
> code and could be adopted for this one as well.

If we're discussing about the input YAML formatting, i.e.:

    instances:
      - jenkins-dev-slave:
          cloud-name: mycloud1
          count: 1
          stop-on-terminate: True
      - jenkins-test-slave:
          cloud-name: mycloud2
          count: 2
          stop-on-terminate: False

v.s.

    instances:
      - cloud-name: mycloud1
        template-name: jenkins-dev-slave
        count: 1
      - cloud-name: mycloud2
        template-name: jenkins-test-slave
        manual-template: True
        count: 2

then I am not convinced that former YAML format (which current JJB jclouds
plugin adopts) is better (although it may make the code look cleaner).

1: Order of configuration stanzas differs between YAML and Jenkins GUI

  Former YAML format will require the name of the template ('template-name' in
  latter YAML format) to appear before 'cloud-name', which contradicts the
  order of input fields in Jenkins GUI (http://pasteboard.co/rDYMTUW.png).

2: 'cloud-name' is not an optional parameter

  'template-name' is only valid in specified 'cloud-name' domain. In other
  words, unlike 'count', 'stop-on-terminate', or 'manual-template' parameter,
  'cloud-name' is not an optional parameter and has an equal significance level
  as 'template-name' parameter.  Putting 'template-name' in shallower indent
  level may make people to think otherwise.

3: Name of the template is not a pre-defined constant

  'template-name' is a user-defined value. Making 'template-name' a key and not
  a value (like the former does) may make it looks like it is a pre-defined
  constant.

4: Other plugins employ a similar YAML format

  From a partial grepping, I found that at least the following plugins use
  similar YAML format (list of flat dictionary):

  artifact-deployer
  artifact-resolver
  batch-tasks
  cloudformation
  conditional-step
  config-file-provider
  dynamic-choice-scriptler
  dynamic-string-scriptler
  multijob
  trigger-builds

I am not suggesting that we should change the jclouds plugin input YAML format
because it will not be backwards compatible. However, I think this is a good
chance to re-think the input YAML format for new plugins.


> b) The plugin changelog suggests that this code may already be out of date
> with respect to the XML now generated. Which may mean that landing this as is
> means we might have to support a change in XML which usually results in
> checking the version of the plugin installed and then branching to the
> different XML formats.

I did check the XML generated by this commit against Jenkins OpenStack Cloud
Plugin version 2.3 (before I pushed the changes, of course) and I could not
find any deviations between XML generated by this commit and XML generated by
upstream plugin. I would appreciate it if you could point me out the
out-of-datedness, I would gladly fix that (even if I had to use
get_plugin_info).

Also, I have checked that it is compatible back to version 2.0, which was
released in 2016-02-02.

> Potentially using the same input layout format as the jclouds plugin would
> allow us to support the old format through a common function and supporting
> the new format as the main content of this module.

I am not convinced about this either, because jclouds plugin is not completely
dead either (last commit to Jenkins JClouds Plugin was 2016-04-04). If we
delegate the main functionality of both JClouds Plugin and OpenStack Cloud
Plugin into one function, that function has to support the future development
of 2 plugins, which surely is not ideal.

Patch-set: 3