Update patch set 1
Abandoned

Closing off for now, doesn't look like the author has time to follow up on it and it's really not the correct way to approach this problem since it breaks existing expected behaviour.

An alternative that would appear to solve the problem the user is experiencing would be the ability to define a setting on a macro that marks it as always requiring expansion so if a user of the macro forgets to pass arguments to it, it will still be expanded by the code.

This would allow for the existing behaviour to be retained, while allowing someone writing macros for reuse to ensure those requiring substitution always get substitution.

Patch-set: 1
Status: abandoned