Add partial support for maven-release-plugin
This patch adds support for the Maven release build wrapper.
The Maven release plugin makes use of the concept of jenkins Actions
to trigger the actual release -
https://wiki.jenkins-ci.org/display/JENKINS/Action+and+its+family+of+subtypes

The actual code for the action -
https://github.com/jenkinsci/m2release-plugin/blob/master/
src/main/java/org/jvnet/hudson/plugins/m2release/M2ReleaseAction.java,
seems to put together all the required
parameters and makes a call to scheduleBuild2 and passes an Action object with it.
Afaik, JJB doesn't support the concept of Actions. If the M2ReleaseAction was an
instance of Builder, instead of Action it would work with this. Currently, I don't
see how the actual build (release) can be triggered from JJB config, and it seems
like we can define the build wrapper alone as part of the job definition, and would
still have to trigger the release from the UI here -
https://integration.wikimedia.org/ci/job/analytics-release-test/m2release

Bug: T132175
Change-Id: Ifc979cdb8e40ad31debff04f745649f23f7ef91a