Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Took a bit of work but I found the command line required to reproduce:

 LC_ALL= LANGUAGE= LANG= jenkins-jobs test <test-case-yaml-file>

Which still resulted in recursion with this change. Rebasing on top of the change https://review.openstack.org/265959 and also trying on latest master without this change both worked.

That's not to say that perhaps we shouldn't ensure that the locale is set when wrapping the stream, but that would be for a different reason.

I think we could also set up a set of tests where we're not checking the output for scenarios such as this.

Patch-set: 1
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1