Update patch set 13
Patch Set 13:

(1 comment)

I'm a little bit wary of departing from the existing style of one entry point per plugin.

On the face of it, if you are doing apps for both android/ios I imagine this would make it harder to have a simple macro to target both platforms with the same default settings since it is easier conceptually for people to grasp the idea of using the same macro with different settings, instead of templating macro names in order to reference different macros.

From that perspective I'd be more inclined to favour the ability to specify a single macro that I could provide a setting to that would determine whether it would output XML for the the IOS or Android set up.

Any other views, thoughts?

Patch-set: 13
Label: Verified=0