Update patch set 4
Patch Set 4:

If the XML outputted is identical with the latest plugin, and with earlier ones then that's great, it means that they aren't delegating the configuration part to the jclouds plugin even though they were reusing it for the backend.


Looking at you're suggested input format:

    instances:
      - cloud-name: mycloud1
        template-name: jenkins-dev-slave
        count: 1
      - cloud-name: mycloud2
        template-name: jenkins-test-slave
        manual-template: True
        count: 2

Is it valid to use the same cloud-name twice? Or does that have to be unique? e.g. will it cause an error in the job in trying configure it with the same name twice, or at start up?

In which case we should probably use the format:

    instances:
      mycloud1:
        template-name: jenkins-dev-slave
        count: 1
      mycloud2:
        template-name: jenkins-test-slave
        manual-template: True
        count: 2


There are plenty of plugins where the list items don't have to be unique, which is where the first format above is fine, but where they are actually unique items, the dict format where the ID is the key is a better choice. Just depends.


I missed the fact that the jclouds one was using the template-name rather than the cloud name as the unique identifier.

Indeed looking closely at what you've pasted suggests that the jclouds input format is actually broken and potentially leads to encouraging mistakes, and we should probably add something in to switch the format and convert existing usages (separate work item though).


So I guess one question remaining, does the 'cloud-name' have to be unique?

Patch-set: 4