Update patch set 1
Abandoned

I've turned this into a proper test in https://review.openstack.org/312197, so I think we can abandon this review. The test will ensure that we don't accidentally re-introduce the issue if the encoding is set to None with a smaller number of jobs that this originally used, but still sufficient to trip it.

Patch-set: 1
Status: abandoned