Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

The extra plugins info is just going to cause confusion, just reference the existing ones in the same folder unless you are adding something that actually checks for the newer version of the plugin and does something different in the code.

I think we should also make this plugin version specific behaviour easier to test by looking up a default file in a version specific folder if it exists.

Patch-set: 2
Label: Code-Review=-1