Update patch set 4
Patch Set 4:

> If the XML outputted is identical with the latest plugin, and with
> earlier ones then that's great, it means that they aren't delegating
> the configuration part to the jclouds plugin even though they were
> reusing it for the backend.
Thank you for understanding.

> Looking at you're suggested input format:
>
>     instances:
>       - cloud-name: mycloud1
>         template-name: jenkins-dev-slave
>         count: 1
>       - cloud-name: mycloud2
>         template-name: jenkins-test-slave
>         manual-template: True
>         count: 2
>
> Is it valid to use the same cloud-name twice? Or does that have to be
> unique? e.g. will it cause an error in the job in trying configure it
> with the same name twice, or at start up?  In which case we should
> probably use the format:
>
>     instances:
>       mycloud1:
>         template-name: jenkins-dev-slave
>         count: 1
>       mycloud2:
>         template-name: jenkins-test-slave
>         manual-template: True
>         count: 2
>
> There are plenty of plugins where the list items don't have to be
> unique, which is where the first format above is fine, but where they
> are actually unique items, the dict format where the ID is the key is
> a better choice. Just depends.
Let's say this, this is valid (minimal example) and will launch 5 VMs:

    instances:
      - cloud-name: liberty-cloud
        template-name: liberty-cloud-jenkins-dev-slave
      - cloud-name: liberty-cloud
        template-name: liberty-cloud-jenkins-stg-slave
      - cloud-name: liberty-cloud
        template-name: liberty-cloud-jenkins-prd-slave
      - cloud-name: mitaka-cloud
        template-name: mitaka-cloud-jenkins-dev-slave
      - cloud-name: mitaka-cloud
        template-name: mitaka-cloud-jenkins-stg-slave

Each VM launched by Jenkins OpenStack Plugin is determined by the
combination of `(cloud-name, template-name)`.

`cloud-name` is a name of a set of following information (excerpt):
* Keystone endpoint URL
* Keystone username
* Keystone password
* OpenStack region

`template-name` is a name of a set of following information (excerpt):
* Flavor name
* Image name
* Key pair name
* Availability zone name

Therefore, neither `cloud-name` alone nor `template-name` alone is
meaningful to determine what VM to launch, and only way to make it
meaningful is to make it as a set of information: `(cloud-name,
template-name)`.

> I missed the fact that the jclouds one was using the template-name
> rather than the cloud name as the unique identifier.  Indeed looking
> closely at what you've pasted suggests that the jclouds input format
> is actually broken and potentially leads to encouraging mistakes, and
> we should probably add something in to switch the format and convert
> existing usages (separate work item though).
Yes, that was something that baffled me when I was writing the first
implementation of this patch (because you can have same `template-name`
in different `cloud-name` but you cannot describe this situation in JJB
jclouds plugin).  BTW, if we were to implement this change (which I may
be able to take), it probably will be backward incompatible (technically
we can check the object type (currently is dictionary and if we
implement the change it will be a list) and determine what to do, but).
What is the recommended procedure to implement these type of changes?

> So I guess one question remaining, does the 'cloud-name' have to be
> unique?
I think I've answered all the questions above.  If we can agree with it,
I'll change the patch with your suggestion and push it.

Patch-set: 4
Label: Verified=0